New wifi-commonsys project to add QtiWifiService service.
This is a new wifi-commonsys project used to have a separate service
parallel to wifi service named as QtiWifiService.

This QtiWifiService is getting started at boot completion of the device.
This service can be used to have oem specific feature development.
Currently this service is being used to collect CSI data from cfrtool via
hidl and then pass the data to application(user level).